fix: Use private subnet for K8s nodes
The Kubernetes nodes do not need to be reachable via the Internet so
should be launched in the private subnets for security.

Also disable auto-public IP assignment for resources in the public
subnet. Resources that actually need public facing IP address should
explicitly request it rather than relying on auto-assignment.

This resolves #56.

Issue: #56
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>