ci: Scale down green node group to minimum 1 node
Currently the "green" node group is configured to require 4 nodes with a max
of 20. We are likely overprovisioning that nodegroup since the services we run
in there aren't all that resource intensive. Let's set the minimum to 1 and
allow to to auto-scale and adjust on it's own.

This closes #78.

Issue: #78
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>