Set resource limits for dind container
Relates to pytorch/pytorch#124727

This sets the cpu/memory resource limits to apply the same limit
as the runner container to ensure that the dind container does not
exceed the limits we set.

Note that we are purposely not also setting a "requests" configuration
to ensure that we are not requesting more cpu/memory resources than
what the runner container is already requesting.

Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>