Limit cpu usage on docker-in-docker containers (take 2) (#180)
Addresses https://github.com/pytorch/ci-infra/issues/94

When the release container launches a container, that container runs in
the dind (docker-in-docker) host container running on the node.

Adding resource limits on CPU, Memory, and GPU usage to the dind
container so that containers executed inside it don't take over the
entire node.

This approach should be okay, since when the runner launches a container
it just waits until the container exits, so the runner won't be using
CPU at the same time as the docker container. However, it would still be
holding onto some memory, and it would be ideal if we could instead
declare a limit at the pod level instead of specifying per-container
limits.

Caveat: While this fix does prevent pods from using more cpus than is
allocated to them, it doesn't prevent pods from _thinking_ that they
have the whole node availabe. `nproc` still returns the full number of
cores on the host instead of the desired "limit" amount

Our system doesn't like fractional CPUs, so using the `m` units instead.
1000m = 1 cpu