Grant runners write access to S3 ossci-benchmarks bucket (#296)
TSIA, this grants the runner the permission to upload benchmark results
to this bucket. Once there, the results will be propagated to
ClickHouse.

The other PR on pytorch-gha-infra
https://github.com/pytorch-labs/pytorch-gha-infra/pull/533. I assume
that I need to have this one to grant the permission to LF runners, plz
correct me if that's not the case.