Reduce minimum runners to 4 (#334)
Back in September we increased this from 0 to 6 to mitigate issues for
runner types that were not frequently used where the Autoscaler dropped
scale-up requests resulting in no runners being available for jobs. At
time we were also testing LF runners at 95% - 100% load balance which is
no longer the case.

Let's reduce the min_available_runners again to try to reduce idle costs
on the account. Lets try 4 for now and reduce it further if no issues
reported.

Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>