[Take 2] Update Terrafile to version used by Meta fleet (#353)
Includes a bunch of reliability fixes from:

* https://github.com/pytorch/test-infra/pull/6934
* https://github.com/pytorch/test-infra/pull/6885
* https://github.com/pytorch/test-infra/pull/6933
* https://github.com/pytorch/test-infra/pull/6932

Previously had to revert these changes to mitigate
https://github.com/pytorch/pytorch/issues/159290. It turns out the
expiration policy added by
https://github.com/pytorch/test-infra/pull/6885 required a setting to be
set on AWS to allow advanced SSM parameters (see
[docs](https://docs.aws.amazon.com/systems-manager/latest/userguide/parameter-store-advanced-parameters.html#ps-default-tier)).

That setting had been set on the Meta AWS account but not on the LF
account, causing the outage. It's now been set on the LF account, so
this should be safe to redeploy.

As a follow up task, when creating the ssm parameter we should
explicitly specify the tier that it should have