Install ArgoCD on the ARC cluster (#363)
Deploy ArgoCD on the ARC cluster by using an helm chart in the terraform
helm layer.

There is no ingress set for ArgoCD for now, the service may be accessed
via port-forward.

The deployed ArgoCD outputs are used to feed the configuration of the
ArgoCD provider in the next layer.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>