Add cert-manager and ingress-nginx to ARC (#365)
Add cert-manager to the k8s cluster used for ARC.
Add a ClusterIssuer. The email address required for that is pulled from
a GitHub secret called CERT_MANAGER_EMAIL that needs to be created.
GitHub workflows have been updated accordingly.

Add ingress-nginx to the k8s cluster used for ARC. This ingress works
nicely with cert-manager, ArgoCD and the AWS LB infra.

Configure ArgoCD to use an ingress with the domain name
argocd.pytorch.org, annotated for cert-manager and nginx. ArgoCD prefers
SSL passthrough, because it transports HTTPs and grpc on the same port.
Using SSL passthrough also enhances security as the traffic stays
encrypted within the cluster as well.

Using SSL passthrough is a bit of an headache for the HTTP01 challenge.
Adding a dedicated ingress template to the cluster issuer should solve
that be creating a dedicated route in the ingress backed by the
challenge pod and with SSL pass-through disabled.

The DNS name needs to be provisioned manually, so the ingress won't have
a valid certificate initially, since Let's Encrypt challange will fail.
Once the DNS record is provisioned, the certificate will be created.

Order of execution (after the PR is reviewed):
- merge this PR
- check the public IP address associated with the new ingress
- provision the DNS record
- verify that the certificate is provisioned correctly

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>