Add the AWS LB controller (#367)
Add the resource required provision the AWS LB controller. This
controller requires a special policy/role to provision an AWS network
load balancer on the fly when an Ingress of the right class is created.

This PR adapts previous resources as well to work with the AWS LB and
fixes an issue with the ClusterIssuer by moving its provisioning to helm
layer to address a limitation of the kubernetes_manifest terraform
module.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>