Remove leading space from provider_path
The provider_path for ARC runner scale sets included an
unwanted leading space, causing path match to fail.

In the arc-deploy-prod workflow, pass TF_VAR_git_revision
to the make apply, so that the git_version var in terraform
is consistent with the branch that the workflow is being
executed from. This allows for testing the ApplicationSet
git directory matrix from the branch as well.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>